requirements.txt
fastapi
uvicorn
langchain
groq
requests
pydantic
python-dotenv

